+++ app/src/main/java/com/ae/apps/tripmeter/listeners/ExpenseChangeListener.java
@@ -1,0 +1,14 @@
+package com.ae.apps.tripmeter.listeners;
+
+/**
+ * Interface that must be implemented to listen to Expense Changes
+ */
+
+public interface ExpenseChangeListener {
+
+    /**
+     * Invioked when an expense is added or removed
+     */
+    public void onExpenseChanged();
+
+}